CAFSC10.20GS/requirements.txt
torch>=2.0
torchvision>=0.15
numpy
scikit-learn
pillow
matplotlib
seaborn
tqdm
